feat(er): remove triage from ER bed workflow
