+++ src/features/bed-dashboard/__tests__/auditor-history-actions.test.ts
@@ -35,8 +35,8 @@ describe('auditor-history-actions', () => {
           id: 'log-1',
           bedId: 'bed-1',
           bedNumber: 'ER-01',
-          fromStageName: 'Triage',
-          toStageName: 'Doctor Assessment',
+          fromStageName: 'Initial Investigation',
+          toStageName: 'Initial Treatment',
           changedByUserId: 'u-1',
           changedByUsername: 'admin1',
           transitionTime: new Date('2026-02-19T08:00:00.000Z'),
@@ -77,8 +77,8 @@ describe('auditor-history-actions', () => {
             id: 'log-1',
             bedId: 'bed-1',
             bedNumber: 'ER-01',
-            fromStageName: 'Triage',
-            toStageName: 'Doctor Assessment',
+            fromStageName: 'Initial Investigation',
+            toStageName: 'Initial Treatment',
             changedByUserId: 'u-1',
             changedByUsername: 'admin1',
             transitionTime: new Date('2026-02-19T08:00:00.000Z'),
@@ -96,8 +96,8 @@ describe('auditor-history-actions', () => {
             id: 'log-2',
             bedId: 'bed-2',
             bedNumber: 'ER-02',
-            fromStageName: 'Registration',
-            toStageName: 'Doctor Assessment',
+            fromStageName: 'Initial Treatment',
+            toStageName: 'Drugs/Test',
             changedByUserId: 'u-2',
             changedByUsername: 'admin2',
             transitionTime: new Date('2026-02-19T09:00:00.000Z'),
