+++ src/features/bed-dashboard/__tests__/auditor-history-queries-archive.test.ts
@@ -27,8 +27,8 @@ describe('fetchAuditorHistory archive coverage', () => {
             id: 'log-1',
             bedId: 'bed-1',
             bedNumber: 'ER-01',
-            fromStageName: 'Triage',
-            toStageName: 'Doctor Assessment',
+            fromStageName: 'Initial Investigation',
+            toStageName: 'Initial Treatment',
             changedByUserId: 'u-1',
             changedByUsername: 'nurse1',
             transitionTime: new Date('2026-02-20T10:00:00.000Z'),
