+++ src/features/bed-dashboard/__tests__/auditor-history-view-readonly.test.tsx
@@ -24,8 +24,8 @@ describe('AuditorHistoryView audit mode read access', () => {
             id: 'log-1',
             bedId: 'bed-1',
             bedNumber: 'ER-01',
-            fromStageName: 'Registration',
-            toStageName: 'Triage',
+            fromStageName: 'Initial Investigation',
+            toStageName: 'Initial Treatment',
             changedByUserId: 'u-1',
             changedByUsername: 'admin1',
             transitionTime: new Date('2026-02-20T10:00:00.000Z'),
