+++ src/features/bed-dashboard/__tests__/transition-queries-archive.test.ts
@@ -25,8 +25,8 @@ describe('getStageTransitions archive coverage', () => {
           id: 'log-1',
           bedNumber: 'ER-01',
           bedId: 'bed-1',
-          fromStageName: 'Triage',
-          toStageName: 'Doctor Assessment',
+          fromStageName: 'Initial Investigation',
+          toStageName: 'Initial Treatment',
           transitionTime: new Date('2026-02-20T10:00:00.000Z'),
           durationInPreviousStageMs: 120000,
           durationInCurrentStageMs: 60000,
