+++ src/features/bed-dashboard/components/BedStageContextMenu.tsx
@@ -21,7 +21,6 @@ interface BedStageContextMenuProps {
   error?: string | null // Error message when loading stages fails
   onStageSelect: (bedId: string, stageId: string) => void
   onClose: () => void
-  onOpenTriage?: (bedId: string) => void // US-20.2
 }
 
 export function BedStageContextMenu({
@@ -38,7 +37,6 @@ export function BedStageContextMenu({
   error,
   onStageSelect,
   onClose,
-  onOpenTriage,
 }: BedStageContextMenuProps) {
   const getActionLabel = (fromStageName: string, toStageName: string) => {
     if (fromStageName === 'Discharge Process' && toStageName === 'Cleaning') {
@@ -90,20 +88,6 @@ export function BedStageContextMenu({
     })
   }, [bed, stages, isMutating, isLoadingTransitions, updatingStageId, isOffline, validNextStages, overrideRequiredStages, onStageSelect])
 
-  // US-20.2: Inject the "Update Triage Details" option at the top if the bed is in Triage (or occupied)
-  const fullItems = useMemo(() => {
-    const result = [...items]
-    if (bed?.currentStage?.name === 'Triage' && onOpenTriage) {
-      result.unshift({
-        id: 'update-triage',
-        label: 'Update Triage Details',
-        onSelect: () => onOpenTriage(bed.id),
-        className: 'text-blue-600 dark:text-blue-400 font-semibold mb-2 border-b pb-2 rounded-none' // stand out
-      })
-    }
-    return result
-  }, [items, bed?.currentStage?.name, bed?.id, onOpenTriage])
-
   if (!bed) {
     return null
   }
@@ -126,7 +110,7 @@ export function BedStageContextMenu({
     <ContextMenu
       isOpen={isOpen}
       position={position}
-      items={error ? [] : fullItems}
+      items={error ? [] : items}
       onClose={onClose}
       header={bed ? `Update ${bed.bedNumber}` : "Update Bed"}
       error={error}
